DTLS 1.3: As a Server, we can complete handshake with PSK.

+++ TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/message/AcknowledgementMessage.java
@@ -64,7 +64,7 @@ public class AcknowledgementMessage extends ProtocolMessage {
         this.parsedRecordNumbers = parsedRecordNumbers;
     }
 
-    public void addRecordNumber(int epoch, long sequenceNumber) {
+    public void addRecordNumber(long epoch, long sequenceNumber) {
         this.parsedRecordNumbers.add(new RecordNumberStruct(epoch, sequenceNumber));
     }
